HopField/InfoField clean up
