+++ src/main/java/org/scion/internal/ScionHeaderParser.java
@@ -19,6 +19,7 @@ import static org.scion.internal.ByteUtil.*;
 import java.net.*;
 import java.nio.ByteBuffer;
 
+import com.google.protobuf.ByteString;
 import org.scion.ScionException;
 import org.scion.ScionPath;
 import org.scion.ScionSocketAddress;
@@ -150,7 +151,7 @@ public class ScionHeaderParser {
         // raw path
         byte[] path = new byte[pos + hdrLenBytes - data.position()];
         data.get(path);
-        PathHeaderScionParser.reversePathInPlace(ByteBuffer.wrap(path));
+        reversePathInPlace(ByteBuffer.wrap(path));
 
         // get remote port from UDP overlay
         data.position(pos + hdrLenBytes);
@@ -277,5 +278,71 @@ public class ScionHeaderParser {
         // Path header
         len += pathHeaderLength;
         return len;
+    }
+
+    public static void reversePathInPlace(ByteBuffer data) {
+        int pos = data.position();
+        int i0 = data.getInt();
+        int seg0Len = readInt(i0, 14, 6);
+        int seg1Len = readInt(i0, 20, 6);
+        int seg2Len = readInt(i0, 26, 6);
+
+        // set CurrINF=0; CurrHF=0; RSV=0
+        int i0R = 0;
+        i0R = writeInt(i0R, 0, 2, 0); // CurrINF = 0
+        i0R = writeInt(i0R, 2, 6, 0); // CurrHF = 0
+        i0R = writeInt(i0R, 8, 6, 0); // RSV = 0
+        // reverse segLen
+        int seg0LenR = seg2Len > 0 ? seg2Len : seg1Len > 0 ? seg1Len : seg0Len;
+        int seg1LenR = seg2Len > 0 ? seg1Len : seg1Len > 0 ? seg0Len : seg1Len;
+        int seg2LenR = seg2Len > 0 ? seg0Len : seg2Len;
+        i0R = writeInt(i0R, 14, 6, seg0LenR);
+        i0R = writeInt(i0R, 20, 6, seg1LenR);
+        i0R = writeInt(i0R, 26, 6, seg2LenR);
+        data.putInt(pos, i0R);
+
+        // info fields
+        int posInfo = data.position();
+        long info0 = data.getLong();
+        long info1 = seg1Len > 0 ? data.getLong() : 0;
+        long info2 = seg2Len > 0 ? data.getLong() : 0;
+        long info0R = seg2Len > 0 ? info2 : seg1Len > 0 ? info1 : info0;
+        long info1R = seg2Len > 0 ? info1 : seg1Len > 0 ? info0 : info1;
+        long info2R = seg2Len > 0 ? info0 : info2;
+
+        data.position(posInfo);
+        long currDirMask = 1L << 56; // For inverting the CurrDir flag.
+        data.putLong(info0R ^ currDirMask);
+        if (seg1LenR > 0) {
+            data.putLong(info1R ^ currDirMask);
+        }
+        if (seg1LenR > 0) {
+            data.putLong(info2R ^ currDirMask);
+        }
+
+        // hop fields
+        int posHops = data.position();
+        int nHops = seg0Len + seg1Len + seg2Len;
+        for (int i = 0, j = nHops - 1; i < j; i++, j--) {
+            int posI = posHops + i * 3 * 4; // 3 int per HopField and 4 bytes per int
+            int posJ = posHops + j * 3 * 4;
+            for (int x = 0; x < 3; x++) {
+                int dummy = data.getInt(posI + x * 4);
+                data.putInt(posI + x * 4, data.getInt(posJ + x * 4));
+                data.putInt(posJ + x * 4, dummy);
+            }
+        }
+
+        data.position(pos);
+    }
+
+    public static void writePath(ByteBuffer data, ByteString path) {
+        for (int i = 0; i < path.size(); i++) {
+            data.put(path.byteAt(i));
+        }
+    }
+
+    public static void writePath(ByteBuffer data, byte[] path) {
+        data.put(path);
     }
 }
